Fix type on IMAGE_GATHER4_C_LZ
Had a different set of types compared to the other IMAGE_GATHER4 ops.

+++ src/shader_recompiler/frontend/format.cpp
@@ -3603,7 +3603,7 @@ constexpr std::array<InstFormat, 112> InstructionFormatMIMG = {{
      ScalarType::Undefined},
     // 79 = IMAGE_GATHER4_C_LZ
     {InstClass::VectorMemImgSmp, InstCategory::VectorMemory, 4, 1, ScalarType::Uint32,
-     ScalarType::Uint32},
+     ScalarType::Float32},
     // 80 = IMAGE_GATHER4_O
     {InstClass::VectorMemImgSmp, InstCategory::VectorMemory, 4, 1, ScalarType::Uint32,
      ScalarType::Float32},
